fix: circular dep
